requirements.txt
# pip install -U -r requirements.txt

flickrapi
numpy
pillow
tqdm
requests
pandas~=1.2.1
pymongo~=3.11.2
exifread
cryptography
streamlit
dnspython
